Use assert_receive/1 to capture emitted events
Telemetry events are sent to the TelemetryCollector GenServer
asynchronously using (GenServer.cast/2).

This means that messages are not guaranteed to be present in the state of
the said process when a subsequent fetch is made
(using GenServer.call(pid, :get_events)).

Instead, we can send events to the current test process and assert that
they are received within a timeout period using assert_receive/1.